Title

Find the posterior distribution of the value of a predictor variable at a particular point along the derivative. For instance, one might be interested in knowing the age value at which a slope is steepest, or closest to zero.

get_X_at_slope_point(
  object,
  term,
  n,
  eps,
  find = c("min", "max", "value"),
  value = NULL,
  multimodal = TRUE,
  prob = 0.95,
  summary_only = FALSE
)

Arguments

object
a brmsfit object

term
character; the smooth term for which the derivative is required.

n
numeric; the number of points to evaluate the derivative at.

eps
numeric; the finite difference.

find
the spot along the derivative to find; can be 'min', 'max', or 'value', in which case the value parameter must be set.

value
if find is set to 'value', this must be set to a specific value, like 0.

multimodal
if the expected posterior might be multimodal, set this to TRUE so that the summary is a true, possibly discontinuous, highest posterior density interval.

prob
the proportion of mass within the HPDI

summary_only
set this to FALSE if you want all samples returned.

Value

either a summary of the posterior density, or a list containing the summary and the posterior samples.